done i hope

+++ includes/Values.hpp
@@ -18,16 +18,16 @@
 
 class Values{
 	protected:
-		int							        _clientMaxBodySize;
-		std::string							_autoIndex;
-		std::string                         _host;
-		std::string							_root;
-		std::string							_serverName;
-		std::string							_upload;
-		std::vector<int>					_ports;
-		std::map<int, std::string>			_errorPages;
-		std::vector<std::string>			_allowMethods;
-		std::vector<std::string>			_indexs;
+		size_t						_clientMaxBodySize;
+		std::string					_autoIndex;
+		std::string					_host;
+		std::string					_root;
+		std::string					_serverName;
+		std::string					_upload;
+		std::vector<int>			_ports;
+		std::map<int, std::string>	_errorPages;
+		std::vector<std::string>	_allowMethods;
+		std::vector<std::string>	_indexs;
 	public:
 		Values();
 		virtual ~Values();
@@ -46,7 +46,7 @@ class Values{
 		void setIndexs(std::string value);
 		// getters
 		std::string					getAutoIndex();
-		int							getClientMaxBodySize();
+		size_t							getClientMaxBodySize();
 		std::string					getHost();
 		std::string					getRoot();
 		std::string					getServerName();
